+++ dataforged-tools/src/classes/oracles/OracleDisplay.ts
@@ -10,19 +10,23 @@ import { cloneDeep } from "lodash-es";
 export class OracleDisplay extends Display implements IOracleDisplay {
   $id: string;
   "Column of"?: IOracle["$id"] | undefined;
-  Columns: (TableColumnText| TableColumnRoll)[];
+  Columns: [TableColumnRoll, ...(TableColumnRoll | TableColumnText)[]];
   constructor(json: IOracleDisplayYaml, parent: IOracle) {
     super(json);
     this.$id = parent.$id + "/Display";
     this["Column of"] = (json["Column of"]) ?? undefined;
     const defaultColumns: (ITableColumnTextYaml | ITableColumnRollYaml)[] = cloneDeep(json.Columns) ?? [{ Type: TableColumnType.Range }, { Type: TableColumnType.String, Key: "Result" }];
     this.Columns = defaultColumns.map((col, index) => {
+      if (index === 0 && col.Type !== TableColumnType.Range) {
+        throw new Error(`${parent.$id} doesn't have a roll column as its first column: ${JSON.stringify(defaultColumns)}`);
+      }
       switch (col.Type) {
         case TableColumnType.Range:
-          return new TableColumnRoll(this.$id, col["Use content from"] ?? parent.$id, index, col.Label);
+          return new TableColumnRoll(this.$id, col["Content"] ?? parent.$id, index, col.Label);
         case TableColumnType.String:
-          return new TableColumnText(this.$id, col["Use content from"] ?? parent.$id, index, col.Label, col.Key);
+          return new TableColumnText(this.$id, col["Content"] ?? parent.$id, index, col.Label, col.Key);
       }
-    });
+    }) as typeof this["Columns"];
+    if (this.Columns.length !== new Set(this.Columns.map(col => col.Label)).size) {throw new Error(`${parent.$id}'s column labels aren't unique ${JSON.stringify(this.Columns)}`);}
   }
 }