+++ dataforged-tools/src/classes/oracles/TableColumn.ts
@@ -8,12 +8,12 @@ export class TableColumnText implements ITableColumnText {
   readonly Type = TableColumnType.String;
   $id: string;
   Label: ITableColumnText["Label"];
-  ["Use content from"]: ITableColumnText["Use content from"];
+  ["Content"]: ITableColumnText["Content"];
   Key: ITableColumnText["Key"];
   constructor(parentID: string, content: IOracle["$id"], index: number, label: string = "Result", key: ITableColumnText["Key"] = "Result") {
     this.$id = `${parentID}/Columns/${index+1}`;
     this.Label = label;
-    this["Use content from"] = content;
+    this["Content"] = content;
     this.Key = key;
   }
 }
@@ -25,10 +25,10 @@ export class TableColumnRoll implements ITableColumnRoll {
   readonly Type = TableColumnType.Range;
   $id: string;
   Label: string;
-  ["Use content from"]: ITableColumnRoll["Use content from"];
+  ["Content"]: ITableColumnRoll["Content"];
   constructor(parentID: string, content: IOracle["$id"], index: number, label: string = "Roll") {
     this.$id = `${parentID}/Columns/${index+1}`;
     this.Label = label;
-    this["Use content from"] = content;
+    this["Content"] = content;
   }
 }