+++ dataforged-tools/src/utils/types/KeysWithValuesOfType.ts
@@ -2,5 +2,5 @@
  * @public
  */
 export type KeysWithValuesOfType<T, V> = keyof {
-  [P in keyof Required<T> as Required<T>[P] extends V ? P : never]: P;
+  [P in keyof T]: T[P] extends V | undefined ? V : never;
 };